+++ activejob/test/cases/adapter_test.rb
@@ -33,4 +33,31 @@ class AdapterTest < ActiveSupport::TestCase
       end
     end
   end
+
+  if adapter_is?(:sneakers)
+    test "sneakers adapter should be deprecated" do
+      before_adapter = ActiveJob::Base.queue_adapter
+
+      msg = <<~MSG.squish
+        The built-in `sneakers` adapter is deprecated and will be removed in Rails 8.1.
+        Please migrate from `sneakers` gem to `kicks` gem version 3.1.1 or later to use `ActiveJob` adapter from `kicks`.
+      MSG
+      assert_deprecated(msg, ActiveJob.deprecator) do
+        ActiveJob::Base.queue_adapter = :sneakers
+      end
+
+    ensure
+      ActiveJob::Base.queue_adapter = before_adapter
+    end
+
+    test "sneakers check_adapter should warn" do
+      msg = <<~MSG.squish
+        The built-in `sneakers` adapter is deprecated and will be removed in Rails 8.1.
+        Please migrate from `sneakers` gem to `kicks` gem version 3.1.1 or later to use `ActiveJob` adapter from `kicks`.
+      MSG
+      assert_deprecated(msg, ActiveJob.deprecator) do
+        ActiveJob::Base.queue_adapter.check_adapter
+      end
+    end
+  end
 end