+++ activejob/CHANGELOG.md
@@ -1,2 +1,7 @@
+*   Deprecate `sneakers` as an adapter option.
+
+    If you're using this adapter, migrate from `snekers` to `kicks` gem version `3.1.1` or later to continue using the adapter.
+
+    *Dino Maric*
 
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/activejob/CHANGELOG.md) for previous changes.